Move accessibility score filter into sandbox
