+++ src/main/java/org/opentripplanner/routing/api/request/ItineraryFilterParameters.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.routing.api.request;
 
 import java.util.function.DoubleFunction;
+import org.opentripplanner.ext.accessibilityscore.AccessibilityScoreFilter;
 import org.opentripplanner.routing.algorithm.filterchain.ItineraryListFilterChainBuilder;
-import org.opentripplanner.routing.algorithm.filterchain.filter.AccessibilityScoreFilter;
 
 /**
  * Group by Similarity filter parameters
